finish check-in
