+++ app/src/main/res/layout/activity_share.xml
@@ -26,6 +26,8 @@
             android:id="@+id/etName"
             android:hint="@string/et_name_fragment"
             android:paddingStart="15dp"
+            android:inputType="text"
+            android:imeOptions="actionNext"
             android:background="@drawable/form_edit_text"
             android:padding="5dp"
             android:layout_width="match_parent"
@@ -41,6 +43,7 @@
             android:id="@+id/etEmail"
             android:hint="@string/et_email_fragment"
             android:paddingStart="15dp"
+            android:inputType="textEmailAddress"
             android:background="@drawable/form_edit_text"
             android:padding="5dp"
             android:layout_width="match_parent"
@@ -53,6 +56,7 @@
             android:layout_marginEnd="20dp"/>
 
         <Button
+            android:id="@+id/btnConfirm"
             android:text="@string/btn_confirm"
             android:background="@drawable/shape_button_confirm"
             android:textAllCaps="false"
@@ -62,5 +66,22 @@
             android:layout_marginTop="10dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"/>
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/clLoaderCheckIn"
+            android:visibility="gone"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:background="@color/transparentLoader">
+
+                <ProgressBar
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"/>
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
